+++ src/exactly_lib/impls/types/expression/descriptions/alias.py
@@ -1,28 +1,45 @@
-from typing import Sequence
+from typing import Sequence, TypeVar
 
 from exactly_lib.definitions import formatting
 from exactly_lib.impls.types.expression import grammar
 from exactly_lib.util.cli_syntax.elements import argument as a
+from exactly_lib.util.name_and_value import NameAndValue
 from exactly_lib.util.textformat.structure.core import ParagraphItem
 from exactly_lib.util.textformat.textformat_parser import TextParser
 
+T = TypeVar('T')
 
-class Description(grammar.PrimitiveDescriptionWithNameAsInitialSyntaxToken):
+
+def entry(alias_name: str,
+          aliased: NameAndValue[grammar.Primitive[T]]) -> NameAndValue[grammar.Primitive[T]]:
+    return NameAndValue(
+        alias_name,
+        grammar.Primitive(
+            aliased.value.parse_arguments,
+            _Description(aliased.name, aliased.value.description())
+        )
+    )
+
+
+class _Description(grammar.PrimitiveDescription):
     def __init__(self,
-                 aliased: str,
-                 arguments: Sequence[a.ArgumentUsage],
+                 name: str,
+                 description: grammar.PrimitiveDescription,
                  ):
-        self._aliased = aliased
-        self._arguments = arguments
+        self._name = name
+        self._description = description
+
+    def initial_argument(self, name: str) -> a.ArgumentUsage:
+        return self._description.initial_argument(name)
 
     @property
     def argument_usage_list(self) -> Sequence[a.ArgumentUsage]:
-        return self._arguments
+        return self._description.argument_usage_list
 
     @property
     def description_rest(self) -> Sequence[ParagraphItem]:
         tp = TextParser({
-            'aliased': formatting.keyword(self._aliased),
+            'aliased': formatting.keyword(self._name),
         })
         return tp.fnap(_DESCRIPTION)